refactor(SignatureStatus): Change format of SignatureStatus Enum
