+++ src/main/scala/com/particeep/api/models/enums/SignatureStatus.scala
@@ -1,9 +1,8 @@
 package com.particeep.api.models.enums
 
-object SignatureStatus {
+sealed trait SignatureStatus extends Product with Enum
 
-  sealed abstract class SignatureStatus extends Product with Enum
-
+object SignatureStatus extends EnumHelper[SignatureStatus] {
   case object CANCELED           extends SignatureStatus
   case object EXPIRED            extends SignatureStatus
   case object FAILED             extends SignatureStatus
@@ -12,7 +11,5 @@ object SignatureStatus {
   case object WAITING            extends SignatureStatus
   case object COMPLETED          extends SignatureStatus
 
-  object SignatureStatus extends EnumHelper[SignatureStatus] {
-    def values: Set[SignatureStatus] = Set(READY, EXPIRED, COMPLETED, CANCELED, FAILED, PENDING_VALIDATION, WAITING)
-  }
+  def values: Set[SignatureStatus] = Set(READY, EXPIRED, COMPLETED, CANCELED, FAILED, PENDING_VALIDATION, WAITING)
 }
