+++ src/main/scala/com/particeep/api/models/signature/Signature.scala
@@ -5,8 +5,7 @@ import play.api.libs.json.{ JsObject, Json, OFormat, Writes }
 import java.time.OffsetDateTime
 
 import com.particeep.api.core.Formatter
-import com.particeep.api.models.enums.SignatureStatus.SignatureStatus
-import com.particeep.api.models.enums.SignerType
+import com.particeep.api.models.enums.{ SignatureStatus, SignerType }
 
 case class Signature(
   id:                       String                  = "",
